refactor(detector.ts): Autozoom disable/enable will not also check status before/after.
Before the disable/enable/start/stop checked the persistant autozoom on/off setting before and after
the command. No instead the autozoom on/off state is checked. The persistant setting will be changed
regardless of the previous setting set. detector.init() will await on autozoom/status which will
resolve when the blob is loaded on camera.

BREAKING CHANGE: start/stop functions will not also set the persistant setting. Now the blob is
loaded on the camera regardless of the persistant setting, so start/stop will always work.