bugfix: Avoid discovery hang for BASE
The try...catch block inside a promise on sdk#setupDeviceDiscoveryListeners for the ATTACH event should be resolved for both cases when the try succeeds or the catch is invoked. This scenario was initiation when a BASE device was discovered and no transport implementation is supported for it so we ended up in the catch block and were not able to complete the execution of the block due to the promise not being marked as completed. Fix was to just add `resolve()` at the end of the catch block.