refactor: remove IAutozoomControl (use ICnnControl)
To make it more intuitive and easier to read, the iAutozoomControl
interface was deleted and the classes the implemented it are now
implementing the ICnnControl. The specific methods that were available
only on IAutozoomControl are still present but not described in the
interface any longer (there is anyway only one class that had those
extra methods and that is the IQ/Boxfish autozoom controler class).