refactor: Use Logger and HuddlyHex from sdk-interfaces
Update the source code to use the static Logger now defined in the
sdk-interfaces npm module. This had to be refactored out so that the
device-api-* modules could use it as well without having to define a
dependency on the sdk. Same thing applies for the hexadecimal values for
the vid and pid of the huddly camera devices.