refactor: Grpc client init only on Ace.ts
Since the GrpcTransport class now is only doing connection establishing
and tearing down, and since Ace.ts already has an implementation for
doing the same thing when in "DevelopmentMode", the GrpcTransport class
will be converted into a dummy class and the grpc connection maintanance
will be done in Ace.ts. This avoids having to add @huddly/camera-proto
and @grpc/grpc-js dependencies in both modules (sdk and device-api-ip)
which has proven to be problematic.