bugfix: Init & setMode should be backward-compatible
Cannot call setMode on a camera running a firmware where this feature is
not supported (was introduced with the Plaza support). The fix involves
making the options "mode" to defaul to 'NORMAL' which means the camera
is in autozoom (GF) mode by default.

Adding additional tests for testing all the edge cases when updating the
autozoom control options.