fix(ipUpgrader): fix bug where upgrader is unable to use new cpio files
This is meant as a low risk effort to make the upgrader work with the new cpio files used for the ip cameras.
To ensure as little impact as possible for now I propose to add possibilities for also using the new image format in another pr for a different release.

* Added additional timeout

* Using cpio-reader callback to progress also when going through unused files